Sessions and channels
Sessions represent adb connections. Channels are independent readers of
such connections. Each channel provides a complete copy of the adb
output.

The indented use of sessions and channels is:

    create_lokatt_session(...)
    start_lokatt_session(...)

    in multiple threads:
        create_lokatt_channel(...)
        while (read_lokatt_channel(...))
            process information
        destroy_lokatt_channel(...)

    ...

    in some main thread:
        stop_lokatt_session(...)
        destroy_lokatt_session(...)

Caveat: there is no actual communication with adb yet. Instead, a shell
script is used as the data pump. Things are currently pretty hard-coded,
so the script must be in lokatt's current working directory or it won't
be found. This is acceptable until adb is properly integrated.

In the future, channels will be associated with a filter which will
limit the adb output returned to the channel client.