Add process name to lokatt message
The process name provides a less volatile field than the current pid
when filtering logcat lines from a given process. By including it in
lokatt messages there is no need for the end user to look it up manually
when the process starts.

The process name is fetched the same way as Android's implementation of
'top' fetches it: by read /proc/[pid]/cmdline. For Java processes,
Zygote will update cmdline to the Java package name once it has forked
and specialized.