liblokatt: add async_channels for threaded message handling
Opening a async_channel will create a new pthread dedicated to the
channel. All message callbacks will arrive on the new thread.

The changes in demux.c make it possible to close lokatt channels.
Channels where previously only closed when the lokatt_session ended.

The new bridge.h is intended as a high level api on top of the existing
lokatt.h api.

Signed-off-by: Martin Wallgren <martin@wallgren.it>
Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>