Add async_channels for threaded message handling
Opening a new async_channel will create a new pthread dedicated to the
channel. All message callbacks will arrive on the new thread.

The changes in demux.c is to make it possible to close lokatt channels.
They where previously only closed when the lokatt_sessoion ended.

The new bridge.h is intented as a high level api on top of lokatt.h.