scripts: pre-commit hook
Add simple pre-commit hook. The script can be run manually or as a git
hook (preferably by creating a symlink from .git/hooks/pre-commit).

The hook performs the following checks for all file types:

  - Warn about trailing whitespace.

  - Warn about spaces before tabs in the indent of a line.

  - Warn about added lines that contain the word 'FIXME'.

The hook performs the following check for C code:

  - Warn about added lines that exceed a line length of 80.

More checks may be added at a later stage.

The warnings are to be considered as suggestions for improvements. False
positives will occur, and there will be cases where the best option is
to ignore a specific warning (for instance, adding a long URL in a
comment in a C file).

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>
Signed-off-by: Martin Wallgren <martin@wallgren.it>