scripts: pre-commit hook
Add simple pre-commit hook. The script can be run manually or as a git
hook (preferably by creating a symlink from .git/hooks/pre-commit).

The hook performs the following checks:

  - Warn about added lines that contain the word 'FIXME' (all files).

  - Warn about added lines that exceed a line length of 80 (only .c
    and .h files).

More checks may be added at a later stage.

The warnings are to be considered as suggestions for improvements. False
positives will occur, and there will be cases where the best option is
to ignore a specific warning (for instance, adding a long URL in a
comment in a C file).

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>