scripts: envsetup for virtualenv
envsetup script for setting up a development environment for lokatt.

source scripts/envsetup.sh will set T to the root of the lokatt git and
add the function lokatt_setup to the shell.

The idea is to run lokatt_setup every time you start working and it will
create the python virtual environment and keep the dependencies up to
date.

Signed-off-by: Martin Wallgren <martin@wallgren.it>
Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>