scripts: only run pep8 if Python code was changed
The inclusion of pep8 in the pre-commit hook requires the Python virtual
environment to be initialized before any commit can be made. This is an
artificial requirement when work is contained to liblokatt.

Inspect the diff contents before running the pep8, and skip the pep8
validation if the diff does not modify any Python files.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>
Signed-off-by: Martin Wallgren <martin@wallgren.it>