scripts: optionally run tests under GDB
Provide a way to easily launch code under GDB. Example usage:

    $ T=test-dict:single_insert GDB=1 make test

GDB and Valgrind modes are mutually exclusive.

Tests are forked before execution. GDB is initially set to switch to the
child process which allows the code under test to be debugged. To
instead debug the test case itself, set follow-fork-mode to parent.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>
Signed-off-by: Martin Wallgren <martin@wallgren.it>