Stack
Quick and dirty implementation of a stack that can hold any type of
objects as long as all objects are of the same type.

Ownership is transferred to the stack when an item is pushed onto it.
Callers may choose to free everything by calling stack_destroy, or to
reclaim ownership of all items by accessing stack.data.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>
Signed-off-by: Martin Wallgren <martin@wallgren.it>