filter: parser
Add support to evaluate filter expressions. After the input string has
been tokenized, the list of tokens are rearranged in Reverse Polish
Notation. A dummy lokatt message is passed through to check for syntax
errors. The filter is then ready for the caller to use.

This commit concludes the work required to create filters and check if
lokatt messages match.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>
Signed-off-by: Martin Wallgren <martin@wallgren.it>