liblokatt: adb backend
First version of backend that reads from a connected device via adb.
This implementation is a significant improvement from the old adb.c
implementation in that it uses 'exec-out' instead of 'shell' to run
logcat. This means the binary stream will not contain invalid \r
characters, which in turn means reading the stream can be done with
good old read(2).

Known caveats:

  - Requires exactly one connected device.

  - Requires a version of adb that supports the 'exec-out' command
    (introduced in system/core commit 5d9d434e, May 2014).

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>
Signed-off-by: Martin Wallgren <martin@wallgren.it>