liblokatt: reintroduce missing message fields
Re-add the missing lokatt_message fields pid, tid, etc. The level, tag
and text fields are stored in the payload field. Before
lokatt_next_event returns, the level, tag and text fields will point to
the correct offset in the returned lokatt_message object.

This commit breaks the dummy backend. This is deliberate and the dummy
backend will be updated to read from a raw 'logcat -B' file in a future
commit.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>